Unify definition of linear variable classes
