+++ dyninstAPI/CMakeLists.txt
@@ -64,7 +64,6 @@ set(_private_headers
     src/instPoint.h
     src/inst-power.h
     src/inst-x86.h
-    src/LinearVariable.h
     src/linux-aarch64.h
     src/linux.h
     src/linux-power.h
